Add encoder (Rust) and tests

+++ src/rust/iced-x86/src/decoder/handlers_legacy.rs
@@ -1373,7 +1373,7 @@ impl OpCodeHandler_Jx {
 				super::instruction_internal::internal_set_op0_kind(instruction, OpKind::NearBranch16);
 				super::instruction_internal::internal_set_near_branch16(
 					instruction,
-					(decoder.read_u16() as u32).wrapping_add(decoder.current_ip32()),
+					(decoder.read_u16() as u32).wrapping_add(decoder.current_ip32()) as u16 as u32,
 				);
 			} else {
 				debug_assert!(decoder.state.operand_size == OpSize::Size32);
@@ -1381,7 +1381,7 @@ impl OpCodeHandler_Jx {
 				super::instruction_internal::internal_set_op0_kind(instruction, OpKind::NearBranch16);
 				super::instruction_internal::internal_set_near_branch16(
 					instruction,
-					(decoder.read_u32() as u32).wrapping_add(decoder.current_ip32()),
+					(decoder.read_u32() as u32).wrapping_add(decoder.current_ip32()) as u16 as u32,
 				);
 			}
 		}
